feature: auth()によるコンポーネント管理
