+++ README.md
@@ -71,3 +71,8 @@ shincodeさんの解説は、初めてライブラリを触る人からある程
 
 **2026/04/28**
 * GitHubのSigninを作成
+
+**2026/04/09**
+* authフックの使い方を学習
+* これにより、自力でのsession管理を廃止
+* asyncコンポーネントとloading.tsxの併用を学習
