+++ next-app/src/app/dashboard/page.tsx
@@ -1,0 +1,11 @@
+import { auth } from "@/auth";
+import { Introduction } from "@/components";
+import DashBoard from "./Dashboard";
+
+export default async function Page() {
+  const session = await auth();
+
+  if (!session?.user) return <Introduction />
+
+  return <DashBoard />
+}
