+++ next-app/src/app/layout.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import { auth } from "@/auth";
+import { SigninForm } from "@/auth/components";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -17,11 +19,15 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const session = await auth();
+
+  if (!session?.user) return <SigninForm />
+
   return (
     <html
       lang="en"
