+++ next-app/src/app/loading.tsx
@@ -1,0 +1,7 @@
+import { AppLoadingBar } from '@/components'
+
+export default function Loading() {
+  return (
+    <AppLoadingBar />
+  )
+}
