+++ next-app/src/auth/index.ts
@@ -1,3 +1,2 @@
 export * from "./auth"
-export { proxy } from "./proxy"
 export * as components from "./components"
