--- next-app/src/auth/proxy.ts
@@ -1,1 +1,0 @@
-export { auth as proxy } from "./auth";
