+++ next-app/src/middleware.ts
@@ -1,0 +1,5 @@
+export { auth as proxy } from "./auth"
+
+export const config = {
+  matcher: '/((?!api|_next/static|_next/image|favicon.ico).*)',
+}
